+++ CHANGELOG.md
@@ -8,10 +8,17 @@
  - Removed warning for empty sections
  - Today and Tomorrow displayed instead of date
  - App does not crash at first connection to laundry
+ - Var. IOS bugs
+
+### Added
+
+ - Date-picker for timetable
+ - Modal for confirmation
+ - Scroll to active time
 
 ### Changed
 
- - Redesigned login page
+ - Redesign machine name
 
 ## [0.0.3]